+++ INASOFT_3.0/Modelos/Remision.cs
@@ -9,7 +9,7 @@ namespace INASOFT_3._0.Modelos
     public class Remision
     {
         private int id;
-        private int cantidad;
+        private double cantidad;
         private string fecha;
         private string descripcion;
         private string tipo_Remision;
@@ -19,7 +19,7 @@ namespace INASOFT_3._0.Modelos
         private int id_Usuario;
 
         public int Id { get => id; set => id = value; }
-        public int Cantidad { get => cantidad; set => cantidad = value; }
+        public double Cantidad { get => cantidad; set => cantidad = value; }
         public string Fecha { get => fecha; set => fecha = value; }
         public int Id_Producto { get => id_Producto; set => id_Producto = value; }
         public int Id_Entrada { get => id_Entrada; set => id_Entrada = value; }
